+++ HISTORY.md
@@ -1,5 +1,7 @@
 # 0.7.1-alpha
 * Moved version to sep file to fix setup.py error
+* Removed urlencode import
+* Added Explicit Raise for 422 errors with Decoded Urls
 
 # 0.7.0-dev1
 * Feature: Added airtable.get() method to retrieve record